Porting lifecycle tests
