+++ lib/core/module.js
@@ -13,12 +13,13 @@ const ApplicationError = require('../util/application_error');
  */
 function initModule (ravelInstance) {
   const self = this;
+  const name = Metadata.getClassMetaValue(Object.getPrototypeOf(self), '@role', 'name');
   // connect any Ravel lifecycle handlers to the appropriate events
   function connectHandlers (decorator, event) {
     const handlers = Metadata.getClassMeta(Object.getPrototypeOf(self), decorator, Object.create(null));
     for (const f of Object.keys(handlers)) {
       ravelInstance.once(event, (...args) => {
-        self.log.trace(`Invoking ${decorator} ${f}`);
+        ravelInstance.log.trace(`${name}: Invoking ${decorator} ${f}`);
         handlers[f].apply(self, args);
       });
     }
@@ -35,7 +36,7 @@ function initModule (ravelInstance) {
   ravelInstance.once('post listen', () => {
     for (const f of Object.keys(handlers)) {
       intervals.push(setInterval(() => {
-        self.log.trace(`Invoking @interval ${f}`);
+        ravelInstance.log.trace(`${name}: Invoking @interval ${f}`);
         handlers[f].handler.call(self);
       }, handlers[f].interval));
     }
